Initial pass at expressions
Expression structs corresponding to the BNF of the book, along with implementations of display for each component of an expression in line with the book.

Consderations on whether to acheive indirection through boxing or constructing expressions relative to an arena (or something similar) were settled by observing that rustc boxes.