Initial parsing of expressions
Following the book, for the most part.
And, the design of the grammar is quite elegant with respect to implementation of the parser.

Implementation of expressions remains a is rough, and is to be refined at some point.

Elsewhere:

A 'Tokens' types introduced as an alias for Vec<Token>.

A fix for the scanner failing to properly recognise slashes (due to failing to advance the character after identifying an initial slash).

A refined 'scan' method which takes anything which can be seen as a refernce to a string.
Ideally, anything to which the chars method can be called would be used.
Though, attempts failed.

Tests have also been expanded to include sequential scans.

And, various methods made public.