Prototype conversions
The book elegantly handled conversions through the Java's object design.
Things aren't some simple in Rust.

The broad idea is to implement a type heirarchy.
Conversion then amounts to identifying a common higher type and matching both.
Given the current values, this heirarchy is mostly linear and so is quite managable.
And, as there's no sense of comparison between contents of values, this is implemented through a comparison lookup which ignores contents.
Though, the null value is for the moment incomparable to any other value, including other instances of itself.

For the moment the heirarchy is implemented along with permitted conversions to the bool value.