Additional conversion detail
A variety of methods have been added to ease conversions, together with conversions for some binary ops.

Particularly useful are methods which chain evaluations to a specific value and then return the inner value.

Elsewhere, methods for easily constructing expressions have been added.

Though, at the moment error handling is temporary, and typically returns an error which should be revised to something more specific.

Following from the previous commit, the graph of types turns out to be unimportant.
Conversion is primarily determined by the operator, and at least for the moment there's no lifting.