Interpreter struct
Interpretation now happens relative to a struct.
This will be useful for environments, I think, though a direct use is for tests.
As, an interpreter supports writing to anything which implements the Write trait, so interpretations of the ast can be writtent to a buffer.

An example test is included for print statements, and mostly works.

Of some small issue is lifetimes, though for the most part these can be ignored.
Though, in particular, the write destination must live longer than the interpreter and some improvements may be possible to avoid needing to drop the interpreter before inspecting the destination.
Still, for now tests are viable.