Scanner simplifications and fixes
Methods to take characters have been unified, with an argument to specify the count.

With this, all instances of taking a pair of characters should be fixed.
For most of these there was failure to advance after identifying the first character.

And, in additon, taking a string should also be fixed.
As, here, there was failure to take the closing double-quote.