Broad updates to interpretation
Evaluation is now part of interpretation, as in general evaluation requires information about the environment.

Following the book, a very basic environment is implement as a map.
This isn't sufficient to handle scope, but for it's a useful base.

Some code to handle declarations has been implemented.
It's somewhat rough, but as with the env, and ok basis.