Support for empty declarations
I think, at least.

Roughly, the rule for Var matches against certain kinds of expression.
For non-empty declarations this is an assignment.
And, as added, for empty delcaractions this is an identifier.

The above makes use of a new Empty expression to allow unified handling of declarations.
An alterantive is to make use of internal optionals, though the expectation is that an empty expression may be useful for other tasks.

Elsewhere some additional ergonomics.

And, all uses of "null" have been replaced with "nil" for consistency.