Environments generalised
The book creates a singly linked list of environments.

This works. It makes sense. It's nice.

Still, it's also simple to create a stack.

As, the elements are hash maps, or rather addresses to hash maps and so even when quickly burrowing to very narrow scope there won't be too much overhead in comparison.

The environment struct for the moment behaves a little like a map with get/insert methods.