Initial pass at blocks
The book handles blocks by creating a block statement.
Block statements contain a list of statements, and things work nicely.
Still, I wonder if this is a little more complex than required.

For, the opening and closing of blocks (at least for the moment) amount to commands to update the environment.
It so happens that these are always paired, but there's no real need for this.
For example, there's no issue with failing to exit a block, it'd mean only that there was no reason to enter the block as the temporary environment is released.
So, why not handle the opening and closing of blocks in this way?

As implemented, entering and exiting a block are handled separately, with left and right braces handled as keywords.
This passes a basic test.