First pass at conditionals
A few changes:

The idea to use comparisons to raise/lower types has been removed.
I completely overlooked how this conflicts with using native comparisons between values.
Though, the details here likely need to be checked again.

The rule for statements now returns a single statement.
This is to allow for easy handling of conditionals, where so way to optionally get an else statement is important.

Whether or not a value is truthy or falsey is implemented as a method.

Otherwise, conditional statements are implemented as one may expect.
Take the condition, recursively take some statement, the see if there's need to recursively take another statement for the else.
These are then wrapped in a case and truthy/falsey are used to evaluate which statement to go with.