revision to internal structure
src is now structured as a lib with the previous main moved to a repl mod which compiles to a bin.

most files have been moved to an interpreter mod, corresponding to the division of the book.
that is, if completed the lib with also include src for a vm.

this may be further revised

no changes to functionality introduced