for loops
for loops are implemented as in the book.
that is, as indirect constructions of while loops.

a slight issue is the optional intialiser, condition, and increment.
to handle these being optional a wrapper around the expression parser has been added which returns an empty expression if some delimiter is hit.
and, in the case the optional parts of a for loop these delimiters are either semicolons or rightparens.
this amounts to logic running through the error path, but seems preferable to returning an expression and then doing work to check if an error was present, though the alternaive could work.

elsewhere, construction helpers for expressions and statements have been prefixed with `mk_` for some clarity.