Separating the interpreter
Functions aren't particularly easy to handle with a contextual environment.
So, to ease implementation environments are now explicity passed are parameters.

And, while doing this the same has been applied to out destinations.
As, with both changes interpretation and evaluation have no state and do not require borrowing their bundling structs as mutable.

Todo at some later point is, perhaps, a rebundle of these components.

Still, for now, the borrow checker is a little easier to tame.