Environment refinements
Interior mutability.

Environments need to branch.
Specifically, for closures.

For, in the case of closures each function should retain access to the enclosing environment while creating further sub-environments.

The standard Rc<RefCell<_>> seems fine for this.
Strictly, there's no direct need for Rc.
However, some indirection is needed, and using Rc helps avoid the complexity of resolving a specific chain of ownership.

Though, for the most part borrows are used in place of introducing an additional count.

To help ease things a new type EnvHandle is introduces as an alias for Rc<RefCell<Env>>.

Elsewhere, required changes for the above have been made.