Semantic revisions
And some simplification, perhaps…

Values don't have much value.

Generally, values are the semantic counterpart to (basic) expressions.

Though, in the context of an interpreter, it makes more sense to use the syntactic objects themselves as the semantic foundation.

With the above, an environment is then a store of basic expressions.

Though, `basic` here may be a slight misnomer.
For, really what's of interest are irreducible expressions.
As, to the current true, false, string, and numeric expressions lambdas are to be added soon(ish).

Lambdas fit the same idea above, roughly.
Though, really, a more standard representation would be as the extension of some relation.
Still, as that's not quite feasible, the intension of the relation will do, as the extensional reduction implicitly exists.

To put this into code, expressions have been split into basic expressions and everything else.
And values (retermed as assignments for the moment) are mapped one-one.

When complete, literals should also fade away, I expect, at least from the AST.