Initial function evaluation
Things are fairly incomplete as there's no handling of what's returned from a function.

Still, an addition basic expression termed (for now) a lambda has been introduced.

Lambdas store an enclosing env, parameters, and a sequence of statements.
When evaluated, the enclosing env is narrowed and each parameter is assigned some expression.

I think either I've missed something in the book, or the issue'll be returned to, but it seems all functions are stored globally.
And, surely this is a mess for closures…
In any case, the implementation given keeps function declaration within scope.

As, roughly, function declaration is sugar for binding a lambda to a variable.

Hm.