Simple pass at return statements
Though, somethig dedicated to parsing lists of statements should be added.

Still, for the moment interpretation has been expanded to return an optional ExprB.
The optional is a return, or else the last expression evaluated.
The implicit return diverges from the book, though as things develop I'm less clear on the motivations for lox as oppossed to the motivation for working through some code.