An additional structure…
Passing around the intended output destination for print isn't ideal.
As, there are various other system related things which may be added, and could be bundled with the output destination.
An input source, for example.

There are a few ways to go, though as an intermediate step the destination has been replaced with a struct.

In the future, it may be preferable to replace the optional enclosing environment to be either an env or a base.