Resolution foundations
In a split with the book, resolution will be part of parsing.

The is to be done with a temporary environment, passed through parsing in a similar way an environment is passed through interpretation.
In contrast to interpretation, the environment does not develop through loops, closures, etc.

Though, it does not need to, as the information stored for resolution is relative.

The key observation wrt. functions / closures is that the relevant environment when interpreting a function is the environment within which the function was defined, not that within which it was called.
So, while it is possible to define and call functions at arbitrary depths, it is not possible to change the relative distance between a function declaration and its enclosing environment, and so relative depth is preserved.

For the moment, an env is passed through the parser.
Next is storing the relevant information for resolution in expressions.