A moments reflection…
Using identifiers as keys in the assignment map is, quite, a bad idea.
As, offsets are relative, and so either hashing should ignore the offset or a search is required.
Easier, then, to return to Strings as keys, now aliased by `Id`.

While fixing this, a few other things have been tidied.