Initial breaks & loops impl
Implementing breaks is not particularly straightforward as breaks exert non-local influence.

For example, a neat idea is to implement breaks for blocks, allowing early exit from a block in a similar way to a return from a fn.
And, then if breaks could also specify a value, returns would be variants of breaks.
And, breaks can easily be parsed by inspecting the statements of a block as they're interpreted.

While nice, this does not have the intended semantics.
For example:

var x = 1
loop {
     if (x == 0) {
          break;
      }
      x = x - 1;
}

Here, the break would only exit the block used to scope the result of a true condition, and would not extend to the of the loop.

Still, statements in part control flow, and the interpretation of statements can be extended to return flow.
So, statements now return an control / value pair.
And, in the case of a break the control bubbles up until a block is found.
Roughly, an exception, though in the ok result, and ready for easy extension.

With breaks, loops are introduced and while loops reduce to loops by prefixing a test on the condition.
For the moment, for loops reduce to while loops, but a direct reduction may be preferred at some point.

The structure of tests has been updated a little, and some additional tests have been added for the next syntax.

There's very little error handling, and things are quite messy.
Still, this approach seems mostly sensible…