Parser ergonomics
Parsers now include a scanner, to avoid copying over information from scanners to parsers.
And, given this parsers have been extended with a scan method.

Further, parsers now have an internal env, which is used so long as parse is called.

Tests have been updated with these features in mind.