Unified error struct (mostly)
Unification of the error struct from the walker is designed to ease error reporting.
In particular, with the unified structs from previous commits into tree walker, errors when walking can be associated with a particular location, and in turn the location of the error in the source can be pointed to.

At the moment all errors have been merged.
Generally, retaining the type distinction between token / parse / interpretation errors would be nice.
However, I don't think the payoff is worthwhile for this toy.

An exception to unification is env errors, as the environment is indirectly tied to the source.
And, so, env error are wrapped (as before) into the / a more specific error.