Initial attempts to track location for error reporting
Though, working through this the endeavour may be in vain.

Location is nice, though information about the token would also be nice.

Still, implemented is specific error initialisers depending on whether the walking is walking the source or a token.
As, in the case of a token it's the location of the token that's of interest.

In vain, though, as I doubt I have the interest to work through this for a walker and in particular for a toy language with dynamic types, classes, etc.