REPL basics
Reading by line is something of a headache.
As, the parser expects is not set up to handle an incomplete block of source.

For example, if a block is input over a few lines parsing would expect the block to close on the first line, going line by line.

I think the way to handle this kind of situation is to expand the REPL and parsing (or maybe scanning), so that an incomplete parse sets the REPL in an expecting-completion state.
(Though, would a stack of opening delimiters be sufficient?)

Still, this is a little much for the tree walker.

So, for the moment the REPL reads to EOF and then parses / interprets.

As this is a very simple REPL, output is handled by evaluating statements, and so interpretation only happens when a statement has been parsed.