Various tidying of error related things
Also, the source is now stored as a vector of chars, with lines cached by a stack of indicies.
Useful, perhaps, to help isolate where an error occured, and to allow for some markup.
Perhaps.
For the moment all that's implemented is printing the line of which an error happened.

An alternative here was to store lines in a vec.